+++ src/game/xo.rs
@@ -85,7 +85,7 @@ impl Game for XO {
     // Main function of the XO return 1 if player 1 win, 2 if player 2 win, 0 if nothing happened
     // 3 if draw
     // */
-    fn make_a_move(&mut self, player_input: usize) -> i32 {
+    fn make_a_move(&mut self, player_input: usize) -> usize {
         if self.number_of_move == 9 {
             return 3;
         }
